+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Unreleased
 - `frames` keyword for `render` for faster rendering
 
+## v0.5.2
+- Added support for local network live streaming
+
 ## v0.5.1
 - added support for Pluto notebooks
 - add alignment options to latex rendering